# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Indian Postal Department Field Operations API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /articles:
    get:
      operationId: listArticles
      tags: [articles]
      summary: List postal articles
      security:
        - bearerAuth: []
      parameters:
        - name: officeId
          in: query
          schema:
            type: string
        - name: operatorId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ArticleStatus"
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Article list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Article"
    post:
      operationId: createArticle
      tags: [articles]
      summary: Issue a postal article to an operator
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArticleRequest"
      responses:
        "201":
          description: Article issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
  /articles/{id}:
    get:
      operationId: getArticle
      tags: [articles]
      summary: Get article by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Article details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
    put:
      operationId: updateArticle
      tags: [articles]
      summary: Update article delivery status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateArticleRequest"
      responses:
        "200":
          description: Updated article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
  /articles/{id}/photo:
    post:
      operationId: uploadArticlePhoto
      tags: [articles]
      summary: Upload the label photo captured as proof for a photo-scanned article
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                photo:
                  type: string
                  format: binary
              required: [photo]
      responses:
        "200":
          description: Photo attached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
  /articles/scan/{barcode}:
    get:
      operationId: scanArticle
      tags: [articles]
      summary: Look up article by barcode
      security:
        - bearerAuth: []
      parameters:
        - name: barcode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Article found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Visits ────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
...
    ArticleStatus:
      type: string
      enum: [pending, delivered, attempted, returned]
    MailType:
      type: string
      enum: [speed_post, registered_post, parcel, gyan_post, magazine_post, blind_mail, regd_newspaper, ordinary_mail, emo, epost, intimation, other]
...
    Article:
      type: object
      properties:
        id:
          type: string
        barcode:
          type: string
        articleNumber:
          type: string
        addressee:
          type: string
        deliveryAddress:
          type: string
        phone:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/ArticleStatus"
        deliveryReason:
          type: string
          nullable: true
        operatorId:
          type: string
          nullable: true
        officeId:
          type: string
        gpsLat:
          type: number
          nullable: true
        gpsLng:
          type: number
          nullable: true
        deliveredAt:
          type: string
          format: date-time
          nullable: true
        issuedAt:
          type: string
          format: date-time
        requiresSignature:
          type: boolean
        requiresPhoto:
          type: boolean
        isCod:
          type: boolean
        codAmount:
          type: number
          nullable: true
        mailType:
          $ref: "#/components/schemas/MailType"
          nullable: true
        careOf:
          type: string
          nullable: true
        houseNumber:
          type: string
          nullable: true
        subarea:
          type: string
          nullable: true
        area:
          type: string
          nullable: true
        postOffice:
          type: string
          nullable: true
        pincode:
          type: string
          nullable: true
        landmark:
          type: string
          nullable: true
        digipin:
          type: string
          nullable: true
        hasUidaiOnFile:
          type: boolean
          description: Whether a UIDAI/Aadhaar number was captured for this article. The raw number is never returned — only its hash is stored.
        scannedPhotoUrl:
          type: string
          nullable: true
          description: Signed URL of the captured label photo, if this article was created via photo scan.
      required: [id, barcode, articleNumber, addressee, deliveryAddress, status, officeId, issuedAt, requiresSignature, requiresPhoto, isCod, hasUidaiOnFile]
    CreateArticleRequest:
      type: object
      properties:
        barcode:
          type: string
        articleNumber:
          type: string
        addressee:
          type: string
        deliveryAddress:
          type: string
        phone:
          type: string
        operatorId:
          type: string
        officeId:
          type: string
        requiresSignature:
          type: boolean
        requiresPhoto:
          type: boolean
        isCod:
          type: boolean
        codAmount:
          type: number
        mailType:
          $ref: "#/components/schemas/MailType"
        careOf:
          type: string
        houseNumber:
          type: string
        subarea:
          type: string
        area:
          type: string
        postOffice:
          type: string
        pincode:
          type: string
        landmark:
          type: string
        digipin:
          type: string
        uidaiNumber:
          type: string
          description: Raw number, accepted here only — hashed server-side and never stored or echoed back in plaintext.
      required: [barcode, articleNumber, addressee, deliveryAddress, officeId]
    UpdateArticleRequest:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/ArticleStatus"
        deliveryReason:
          type: string
        gpsLat:
          type: number
        gpsLng:
          type: number
        operatorId:
          type: string
        mailType:
          $ref: "#/components/schemas/MailType"
        careOf:
          type: string
        houseNumber:
          type: string
        subarea:
          type: string
        area:
          type: string
        postOffice:
          type: string
        pincode:
          type: string
        landmark:
          type: string
        digipin:
          type: string
        uidaiNumber:
          type: string
          description: Raw number, accepted here only — hashed server-side and never stored or echoed back in plaintext.

    # ── Visit ─────────────────────────────────────────────────────────
...