# packages/api-zod/src/generated/types/articleStatus.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Indian Postal Department Field Operations API
 * OpenAPI spec version: 0.1.0
 */

export type ArticleStatus = typeof ArticleStatus[keyof typeof ArticleStatus];


export const ArticleStatus = {
  pending: 'pending',
  delivered: 'delivered',
  attempted: 'attempted',
  returned: 'returned',
} as const;